Consolidating files to one report/review file
Deleted report_v2, report_v3, and review_v2 after moving most recent code to report and review (respectively). Changed bot.py to pull from correct files.
