+++ DiscordBot/review.py
@@ -3,15 +3,13 @@ from enum import Enum, auto
 import discord
 import re
 import asyncio
-from datetime import datetime
+from datetime import datetime, timedelta
 
 class ReportType(Enum):
-    DOXXING = "Doxxing"
+    FRAUD = "Fraud"
+    INAPPROPRIATE_CONTENT = "Inappropriate Content"
     HARASSMENT = "Harassment"
-    HATE_SPEECH = "Hate Speech"
-    SPAM = "Spam"
-    INAPPROPRIATE = "Inappropriate Content"
-    OTHER = "Other"
+    PRIVACY = "Privacy"
 
 class InfoType(Enum):
     CONTACT = "Contact Information"
@@ -28,7 +26,9 @@ class State(Enum):
     REVIEW_START = auto()
     AWAITING_MESSAGE = auto()
     MESSAGE_IDENTIFIED = auto()
+    # New states for review_v2
     AWAITING_THREAT_JUDGEMENT = auto()
+    AWAITING_OTHER_ABUSE_JUDGEMENT = auto()
     AWAITING_DISALLOWED_INFO = auto()
     AWAITING_CONTENT_CHECK = auto()
     AWAITING_INTENTION = auto()
@@ -45,15 +45,27 @@ class Review:
         self.client = client
         self.report = None
         self.report_details = None
-        self.info_types = []  # Changed from info_type to info_types list
+        self.info_types = []
         self.details = None
         self.reporter_id = None
         self.timestamp = None
-        self.remove = False
-        self.ban_user = False
-        self.second_reviewer = False
         
-    async def handle_message(self, message):
+        # Assessment flags set by the reviewer for building a summary
+        self.threat_identified_by_reviewer = False
+        self.disallowed_info_identified = False 
+        self.other_problematic_content_identified = False 
+
+        # Flag to indicate if the original reported message should be deleted, suspended, permanently banned, and to escalate to a secondary reviewer
+        self.remove = False
+        self.suspend_user = False
+        self.ban_user = False 
+
+        # Stored Discord objects for performing moderation actions- these are fetched based on information in the report_details embed
+        self.original_reported_message = None 
+        self.original_message_author = None
+        self.abuse_type = None
+
+    async def handle_message(self, message: discord.Message):
         '''
         This function handles the reporting flow by managing state transitions
         and prompts at each state.
@@ -67,7 +79,6 @@ class Review:
             return [self.get_help_message()]
             
         if self.state == State.REVIEW_START:
-            # self.reporter_id = message.author.id
             self.timestamp = datetime.now()
             reply = "Thank you for starting the reviewing process. "
             reply += "Say `help` at any time for more information.\n\n"
@@ -77,26 +88,80 @@ class Review:
             return [reply]
             
         if self.state == State.AWAITING_MESSAGE:
-            # Parse out the three ID strings from the message link
-            print("Link received")
-            m = re.search('/(\d+)/(\d+)/(\d+)', message.content)
+            # Parse IDs from linked report message
+            m = re.search(r'/(\d+)/(\d+)/(\d+)', message.content)
             if not m:
-                return ["I'm sorry, I couldn't read that link. Please try again or say `cancel` to cancel."]
+                return ["Invalid message link. Please use a valid Discord message link or type `cancel`."]
             
-            guild = self.client.get_guild(int(m.group(1)))
+            guild_id, channel_id, message_id = map(int, m.groups())
+            guild = self.client.get_guild(guild_id)
             if not guild:
-                return ["I cannot accept reports of messages from guilds that I'm not in. Please have the guild owner add me to the guild and try again."]
-            
-            channel = guild.get_channel(int(m.group(2)))
-            if not channel:
-                return ["It seems this channel was deleted or never existed. Please try again or say `cancel` to cancel."]
+                return ["Error: Bot cannot find the server for that link."]
+            channel = guild.get_channel(channel_id)
+            if not channel or not isinstance(channel, discord.TextChannel):
+                return ["Error: Mod channel not found or not a text channel."]
             
             try:
-                self.report = await channel.fetch_message(int(m.group(3)))
+                # Fetch the report embed itself
+                self.report = await channel.fetch_message(message_id)
+                if not self.report.embeds or not self.report.author.id == self.client.user.id:
+                    return ["Linked message isn't a valid report embed from me. Please link the correct report."]
                 self.report_details = self.report.embeds[0]
-            except discord.errors.NotFound:
-                return ["It seems this message was deleted or never existed. Please try again or say `cancel` to cancel."]
-            
+
+                original_message_link_str, original_author_id_str = None, None
+                # Extract original message data from report fields
+                for field in self.report_details.fields:
+                    if field.name == "**Direct Link to Reported Message**": 
+                        match = re.search(r'\((.*?)\)', field.value) 
+                        if match: original_message_link_str = match.group(1)
+                    elif field.name == "**Author of Reported Message**": 
+                        match = re.search(r'ID: `(\d+)`', field.value)
+                        if match: original_author_id_str = match.group(1)
+                    if field.name == "**Specific Reason Provided by Reporter**":
+                        self.abuse_type = field.value
+                
+                if not original_message_link_str and self.report_details.description:
+                    match = re.search(r'https?://discord\.com/channels/(\d+)/(\d+)/(\d+)', self.report_details.description)
+                    if match: original_message_link_str = match.group(0)
+                        
+                if not original_message_link_str or not original_author_id_str:
+                    return ["Error: Missing original message details in report embed. Format issue?"]
+
+                link_parts = original_message_link_str.strip("<>").split('/')
+                if len(link_parts) < 3:
+                    return [f"Error: Invalid original message link format in report: {original_message_link_str}"]
+                
+                try:
+                    orig_guild_id, orig_channel_id, orig_message_id = map(int, link_parts[-3:])
+                    orig_author_id = int(original_author_id_str)
+                except ValueError:
+                    return ["Error parsing IDs from report embed. Corrupted data?"]
+
+                orig_guild = self.client.get_guild(orig_guild_id)
+                if not orig_guild: return [f"Error: Original guild (ID: {orig_guild_id}) not found."]
+                
+                orig_channel = orig_guild.get_channel(orig_channel_id)
+                if not orig_channel or not isinstance(orig_channel, (discord.TextChannel, discord.Thread, discord.VoiceChannel, discord.ForumChannel)):
+                     return [f"Error: Original channel (ID: {orig_channel_id}) not found/fetchable."]
+                
+                message_fetchable_channel : discord.abc.Messageable = orig_channel
+                try:
+                    # Fetch the actual user message that was reported
+                    self.original_reported_message = await message_fetchable_channel.fetch_message(orig_message_id)
+                except discord.NotFound: return ["Error: Original reported message not found (deleted?)."]
+                except discord.Forbidden: return ["Error: Bot permission issue fetching original message (Need 'Read History')."]
+
+                try:
+                    self.original_message_author = await orig_guild.fetch_member(orig_author_id)
+                except discord.NotFound:
+                    try:
+                        self.original_message_author = await self.client.fetch_user(orig_author_id)
+                    except discord.NotFound: return ["Error: Original author not found (neither member nor user)."]
+                except discord.Forbidden: return ["Error: Bot permission issue fetching author (Need 'Members Intent'?)."]
+
+            except (discord.errors.NotFound, IndexError) as e:
+                return ["Error: Could not process linked report. Deleted or invalid format?"]
+
             self.state = State.AWAITING_THREAT_JUDGEMENT
             
             reply = f"I found this report:\n"
@@ -105,159 +170,201 @@ class Review:
             reply += "Does this message contain a threat of violence? Selecting yes will result in the post being removed.\n"
             reply += "1. Yes\n"
             reply += "2. No"
-            
+                
             return [reply]
         
-        if self.state == State.AWAITING_THREAT_JUDGEMENT:
-            if message.content == "1":
-                self.threat = "Threat reported."
+        elif self.state == State.AWAITING_THREAT_JUDGEMENT:
+            if message.content == "1": 
+                self.threat_identified_by_reviewer = True
                 self.remove = True
-            elif message.content == "2":
-                pass
+                self.suspend_user = True
+            elif message.content == "2": 
+                self.threat_identified_by_reviewer = False
             else:
-                return ["Please enter a 1 for 'Yes' or a 2 for 'No'"]
+                return ["Invalid input. Please type 1 for Yes or 2 for No."]
             
-            self.state = State.AWAITING_DISALLOWED_INFO
-            reply = "Does this post contain any of the following disallowed information:\n"
-            reply += "-Government ID information (e.g. Social Security Numbers, ID numbers, etc.)\n"
-            reply += "-Financial information (e.g. bank account numbers, credit card numbers, etc.)\n"
-            reply += "1. Yes\n"
-            reply += "2. No"
-            return [reply]
-        
-        if self.state == State.AWAITING_DISALLOWED_INFO:
-            if message.content == "1":
+            if self.abuse_type == "Doxxing":
+                self.state = State.AWAITING_DISALLOWED_INFO
+                reply = ("Does the post contain **disallowed information** (Gov ID, Financial Details)?\n1. Yes\n2. No")
+                return [reply]
+            else:
+                self.state = State.AWAITING_OTHER_ABUSE_JUDGEMENT
+                reply = f"I found this report:\n"
+                for field in self.report_details.fields:
+                    reply += f"{field.name}: {field.value}\n"
+                reply += f"This message was flagged for {self.abuse_type}. Should this message be removed on the basis of that content?\n"
+                reply += "1. Yes\n"
+                reply += "2. No"
+                return [reply]
+            
+        elif self.state == State.AWAITING_OTHER_ABUSE_JUDGEMENT:
+            if message.content == "1": 
                 self.remove = True
-                self.state = State.CONFIRMING_REVIEW
-                reply = "Thank you. If you confirm this review, this post will be immediately removed and the user will be suspended.\n"
-                reply += "Would you like to confirm this review?\n"
-                reply += "1. Yes\n"
-                reply += "2. No"
-                return [reply]
-            elif message.content == "2":
-                self.state = State.AWAITING_CONTENT_CHECK
-                reply = "Does this post contain any of the following:\n"
-                reply += "-Phone numbers\n"
-                reply += "-Email addresses\n"
-                reply += "-Name of employer\n"
-                reply += "-Physical location (e.g. home address, hometown, etc.)\n"
-                reply += "1. Yes\n"
-                reply += "2. No"
-                return [reply]
-            else:
-                return ["Please enter a 1 for 'Yes' or a 2 for 'No'"]
-        
-        if self.state == State.AWAITING_CONTENT_CHECK:
-            if message.content == "1":
-                self.state = State.AWAITING_INTENTION
-                reply = "Was this post shared in good faith by either the potentially targeted individual and/or someone who knows the potentially targeted individual?\n"
-                reply += "1. Yes\n"
-                reply += "2. No"
-                return [reply]
-            elif message.content == "2":
-                self.state = State.CONFIRMING_REVIEW
-                if self.remove:
-                    reply = "Thank you. If you confirmt this review, this post will be removed, as you flagged it as a threat.\n"
-                    reply += "Would you like to confirm this review?\n"
-                    reply += "1. Yes\n"
-                    reply += "2. No"
-                else:
-                    reply = "Thank you. If you confirm this review, no action will be taken.\n"
-                    reply += "Would you like to confirm this review?\n"
-                    reply += "1. Yes\n"
-                    reply += "2. No"
-                return [reply]
-            else:
-                return ["Please enter a 1 for 'Yes' or a 2 for 'No'"]
-        
-        if self.state == State.AWAITING_INTENTION:
-            if message.content == "1":
-                self.state = State.CONFIRMING_REVIEW
-                if self.remove:
-                    reply = "Thank you. If you confirm this review, this post will still be removed, as you flagged it as a threat."
-                    reply += "Would you like to confirm this review?\n"
-                    reply += "1. Yes\n"
-                    reply += "2. No"
-                else:
-                    reply = "Thank you. If you confirm this review, no action will be taken."
-                    reply += "Would you like to confirm this review?\n"
-                    reply += "1. Yes\n"
-                    reply += "2. No"
-                return [reply]
-            elif message.content == "2":
-                if self.remove:  # post had been tagged as threat
-                    self.state = State.EVALUATE_SEVERITY
-                    reply = "Thank you. Since you flagged this post as both a threat and as a doxxing attempt, if you confirm this review, it will be passed to a second reviewer who may choose to contact law enforcement."
-                    reply += "Would you like to confirm this review?\n"
-                    reply += "1. Yes\n"
-                    reply += "2. No"
-                else: 
-                    self.remove = True
+            self.state = State.CONFIRMING_REVIEW
+            if self.threat_identified_by_reviewer:
                     self.state = State.CONFIRMING_REVIEW
-                    reply = "Thank you. If you confirm this review, this post will be removed."
-                    reply += "Would you like to confirm this review?\n"
-                    reply += "1. Yes\n"
-                    reply += "2. No"
-                return [reply]
-            else:
-                return ["Please enter a 1 for 'Yes' or a 2 for 'No'"]
-            
-        if self.state == State.CONFIRMING_REVIEW:
+                    reply = ("Threat identified. Policy: Message removal & 1-day user suspension.\n\n" 
+                            "Confirm review and actions?\n1. Yes (Proceed)\n2. No (Cancel Review)")
+            else: 
+                reply = ("Review assessment (no direct threat ID'd by you):\n")
+                if self.remove:
+                    reply += f"- {self.report_details['**Specific Reason Provided by Reporter**']} content was identified.\n"
+                    reply += "- This will be logged. The post will be removed.\n"
+                else: 
+                    reply += "- No direct threat or other significant problematic content was flagged by you.\n"
+                reply += "No suspension will occur (policy requires reviewer to ID direct threat).\n\n"
+                reply += "Finalize and log assessment?\n1. Yes (Finalize)\n2. No (Cancel Review)"
+            return [reply]
+                
+        
+        elif self.state == State.AWAITING_DISALLOWED_INFO:
             if message.content == "1":
-                # Submit the report to moderators
+                self.disallowed_info_identified = True
+            elif message.content == "2":
+                self.disallowed_info_identified = False
+            else:
+                return ["Invalid input. Please type 1 for Yes or 2 for No."]
+            self.state = State.AWAITING_CONTENT_CHECK
+            reply = ("Does it contain **other problematic content** (phone, email, location, employer)?\n1. Yes\n2. No")
+            return [reply]
+        
+        elif self.state == State.AWAITING_CONTENT_CHECK:
+            if message.content == "1":
+                self.other_problematic_content_identified = True
+            elif message.content == "2":
+                self.other_problematic_content_identified = False
+            else:
+                return ["Invalid input. Please type 1 for Yes or 2 for No."]
+            self.state = State.CONFIRMING_REVIEW
+            if self.threat_identified_by_reviewer:
+                 self.state = State.CONFIRMING_REVIEW
+                 reply = ("Threat identified. Policy: Message removal & 1-day user suspension.\n\n" 
+                          "Confirm review and actions?\n1. Yes (Proceed)\n2. No (Cancel Review)")
+            else: 
+                 reply = ("Review assessment (no direct threat ID'd by you):\n")
+                 if self.disallowed_info_identified or self.other_problematic_content_identified:
+                     reply += "- Problematic content (Disallowed Info or Other) was identified.\n"
+                     reply += "- This will be logged. Manual moderator follow-up may be appropriate.\n"
+                 else: 
+                     reply += "- No direct threat or other significant problematic content was flagged by you.\n"
+                 reply += "No suspension will occur (policy requires reviewer to ID direct threat).\n\n"
+                 reply += "Finalize and log assessment?\n1. Yes (Finalize)\n2. No (Cancel Review)"
+            return [reply]
+        
+        elif self.state == State.CONFIRMING_REVIEW:
+            if message.content == "1":
                 await self._submit_report_to_mods()
-                            
-                response = "The result of this review will now be sent to the moderation team for visibility. Thank you again for your work.\n\n"
-                response += "If you need to review another report, type in the reviewing password again."
                 self.state = State.REVIEW_COMPLETE
-
-                return [response]
+                return ["Review finalized. Outcome logged to the moderator channel. Type the password to start a new review."]
             elif message.content == "2":
                 self.state = State.REVIEW_COMPLETE
-                return ["Review cancelled. If you need to submit a different review, type the moderator password again."]
+                return ["Review cancelled as per your request."]
             else:
-                return ["Please type `1` to confirm or `2` to cancel."]
-                # Submit the report to moderators
+                return ["Invalid input. Please type 1 to Confirm or 2 to Cancel."]
         
-        # If we somehow end up in an unknown state
-        return ["An error occurred in the reviewing process. Please type the moderator password to start again."]
+        return ["An error occurred. Please type `cancel` or contact an admin."]
     
+    async def _execute_moderation_actions(self):
+        """
+        Performs automated moderation actions (message deletion, user suspension) 
+        based on the flags set during the review process
+
+        """
+        actions_taken_summary_list = []
+        
+        if not self.original_reported_message or not self.original_message_author:
+            actions_taken_summary_list.append("Critical Error: Original message/author not identified by bot.")
+            return actions_taken_summary_list
+
+        if self.threat_identified_by_reviewer:
+            # Attempt to delete reported message
+            if self.remove:
+                try:
+                    await self.original_reported_message.delete()
+                    actions_taken_summary_list.append("Original message **deleted**.")
+                except discord.Forbidden:
+                    actions_taken_summary_list.append("Bot **failed to delete** message (Permissions error).")
+                except discord.NotFound:
+                    actions_taken_summary_list.append("Original message **not found** (already deleted?).")
+                except discord.HTTPException as e:
+                    actions_taken_summary_list.append(f"Bot **failed to delete** message (HTTP Error: {e.status}).")
+            
+            if self.suspend_user:
+                if isinstance(self.original_message_author, discord.Member):
+                    try:
+                        timeout_duration = timedelta(days=1)
+                        actions_taken_summary_list.append(f"User `{self.original_message_author.display_name}` **would be suspended for 1 day** (TESTING - action disabled).")
+                    except discord.Forbidden:
+                        actions_taken_summary_list.append(f"Bot **failed to suspend** user `{self.original_message_author.display_name}` (Permissions error).")
+                    except discord.HTTPException as e:
+                        actions_taken_summary_list.append(f"Bot **failed to suspend** user `{self.original_message_author.display_name}` (HTTP Error: {e.status}).")
+                else:
+                    actions_taken_summary_list.append(f"User `{self.original_message_author.display_name}` **could not be suspended** (not a current server member).")
+        
+        if not actions_taken_summary_list: 
+             actions_taken_summary_list.append("No automated actions (delete/suspend) triggered (e.g., no direct threat ID'd by reviewer).")
+
+        return actions_taken_summary_list
+
     async def _submit_report_to_mods(self):
         """
         Send the review to the moderator channel for the guild.
         """
-        if not self.report:
+        if not self.report or not self.report_details:
             return
         
-        # Get the mod channel for this guild
-        guild_id = self.report.guild.id
-        if guild_id not in self.client.mod_channels:
-            return
-        
-        mod_channel = self.client.mod_channels[guild_id]
-
-        embed = None
-
-        if self.second_reviewer:
-            embed = discord.Embed(
-                title=f"Second Reviewer Needed: Doxxing, Potential Threat",
-                color=0xe74c3c, # red
-                timestamp=self.timestamp
-            )
-            embed.add_field(name="Original Message", value=self.report_details["Reported Message"])
-            embed.add_field(name="Link", value=f"[Click to view]({self.report_details['Message Link']})", inline=True)
-
+        reviewer_name = "Moderator (Reviewer ID N/A)" 
+        actions_performed_strings = await self._execute_moderation_actions()
+        summary_lines = []
+        summary_lines.append(f"**Review of Report: `{self.report_details.title}`**") 
+        if self.original_reported_message and self.original_message_author:
+            summary_lines.append(f"Original Msg Author: `{self.original_message_author.display_name}` (ID: `{self.original_message_author.id}`)")
+            summary_lines.append(f"Original Msg Link: [Click to View]({self.original_reported_message.jump_url})")
         else:
-            embed = discord.Embed(
-                title=f"Review Completed",
-                color=0x95a5a6, # gray
-                timestamp=self.timestamp
-            )
-            embed.add_field(name="Reviewed Report", value=self.report_details)
+            summary_lines.append("Warning: Original msg/author details unavailable.")
+
+        # Reviewer's assessment
+        summary_lines.append("\n**Moderator's Assessment:**")
+        summary_lines.append(f"- Threat of Violence: `{'Yes' if self.threat_identified_by_reviewer else 'No'}`")
+        if self.abuse_type == "Doxxing":
+            summary_lines.append(f"- Disallowed Info: `{'Yes' if self.disallowed_info_identified else 'No'}`")
+        summary_lines.append(f"- Other Problematic Content: `{'Yes' if self.other_problematic_content_identified else 'No'}`")
+
+        summary_lines.append("\n**Outcome:**")
+        summary_lines.append("- Status: First-Level Review Completed.")
+        summary_lines.append("\n  **Automated Bot Actions:**")
+        if actions_performed_strings:
+            for item in actions_performed_strings: summary_lines.append(f"    - {item}")
+        else: 
+            summary_lines.append("    - No automated actions logged.") 
         
-        # Send the report to the mod channel
-        await mod_channel.send(embed=embed)
-    
+        if (not self.threat_identified_by_reviewer and 
+            (self.disallowed_info_identified or self.other_problematic_content_identified)):
+                 summary_lines.append("\n  Note: Other violations noted. Manual follow-up may be needed.")
+
+        final_summary_text = "\n".join(summary_lines)
+        
+        embed_title = "Review Finalized"
+        embed_color = discord.Color.dark_grey()
+        if self.threat_identified_by_reviewer: 
+            embed_title, embed_color = "Review: Actions Taken/Logged", discord.Color.green()
+        elif (self.disallowed_info_identified or self.other_problematic_content_identified): 
+            embed_title, embed_color = "Review: Findings Logged", discord.Color.blue()
+
+        guild_id = self.report.guild.id
+        mod_channel = self.client.mod_channels.get(guild_id)
+        if not mod_channel:
+            return
+
+        review_embed = discord.Embed(title=embed_title, description=final_summary_text, color=embed_color, timestamp=datetime.now())
+        review_embed.set_footer(text=f"Review by: {reviewer_name} | Bot v2.2")
+        try:
+            await mod_channel.send(embed=review_embed)
+        except discord.Forbidden:
+            pass
+        except discord.HTTPException as e:
+            pass
+
     def get_help_message(self):
         help_msg = "**Discord Report Bot Help**\n\n"
         
@@ -273,6 +380,11 @@ class Review:
 
         elif self.state == State.AWAITING_THREAT_JUDGEMENT:
             help_msg += "Please identify whether the post in question contains a threat of violence. Posts labeled as a threat will be removed.\n\n"
+            help_msg += "1. Yes\n"
+            help_msg += "2. No\n"
+
+        elif self.state == State.AWAITING_OTHER_ABUSE_JUDGEMENT:
+            help_msg += f"The report was labeled as {self.abuse_type}. Does this content contain {self.abuse_type}?\n"
             help_msg += "1. Yes\n"
             help_msg += "2. No\n"
